test paused/stabdby for async replication
