+++ ydb/tests/functional/transfer/ya.make
@@ -16,7 +16,8 @@ PEERDIR(
 )
 
 SRCS(
-    main.cpp
+    replication.cpp
+    transfer.cpp
 )
 
 INCLUDE(${ARCADIA_ROOT}/ydb/public/tools/ydb_recipe/recipe.inc)
